Separating createMaterialPalette into its own module
